refactor: moving RedirectException.
