+++ user_guide_src/source/changelogs/v4.4.0.rst
@@ -143,6 +143,7 @@ Deprecations
   ``ExceptionHandler``.
 - **Autoloader:** ``Autoloader::sanitizeFilename()`` is deprecated.
 - **CodeIgniter:** ``CodeIgniter::$returnResponse`` property is deprecated. No longer used.
+- **RedirectException:** ``\CodeIgniter\Router\Exceptions\RedirectException`` is deprecated. Use \CodeIgniter\HTTP\Exceptions\RedirectException instead.
 
 Bugs Fixed
 **********
